requirements.txt
aiogram>=3.0.0
python-dateutil>=2.8.2
pydantic>=2.0.0
pytz>=2023.3
aiohttp>=3.9.0
